Move PID base headers to Framework
