+++ Analysis/DataModel/include/AnalysisDataModel/PID/TOFReso.h
@@ -21,7 +21,7 @@
 // Root includes
 #include "TMath.h"
 // O2 includes
-#include "AnalysisDataModel/PID/ParamBase.h"
+#include "Framework/ParamBase.h"
 
 namespace o2::pid::tof
 {
