+++ Analysis/DataModel/include/AnalysisDataModel/PID/TPCReso.h
@@ -18,7 +18,7 @@
 #ifndef O2_ANALYSIS_PID_TPCRESO_H_
 #define O2_ANALYSIS_PID_TPCRESO_H_
 
-#include "AnalysisDataModel/PID/ParamBase.h"
+#include "Framework/ParamBase.h"
 
 namespace o2::pid::tpc
 {
